move EvaluationResult and calculate_metrics to schema
